+++ src/serve/firestore.ts
@@ -1,4 +1,0 @@
-import { FirestoreEmulator } from "../emulator/firestoreEmulator";
-import { EmulatorServer } from "../emulator/emulatorServer";
-
-module.exports = new EmulatorServer(new FirestoreEmulator({}));